Increase max regions per room in BOA calculation
Fixes the AOH Prototype mission, which has a room with 195 regions.
This is too much for the previous limit of 200 since the code always
adds 27 lookup entries. So increase the limit to 250.
(https://sectorgame.com/d3/downloads/files/AOH_Prototype.zip)

Also has a fix for the assert that was supposed to catch this.